+++ src/Base/BaseMongoRepository.php
@@ -1,0 +1,34 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Maatify\DataRepository\Base;
+
+use Maatify\Common\Contracts\Adapter\AdapterInterface;
+use MongoDB\Collection;
+use MongoDB\Database;
+
+abstract class BaseMongoRepository extends BaseRepository
+{
+    public function __construct(AdapterInterface $adapter)
+    {
+        parent::__construct($adapter);
+    }
+
+    protected function getDatabase(): Database
+    {
+        $driver = $this->assertDriverAvailable(
+            $this->adapter->getDriver(),
+            'MongoDB Database'
+        );
+
+        return $this->assertDriverInstance($driver, Database::class);
+    }
+
+    protected function getCollection(string $name): Collection
+    {
+        $database = $this->getDatabase();
+
+        return $database->selectCollection($name);
+    }
+}