+++ src/Base/BaseMySQLRepository.php
@@ -1,0 +1,37 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Maatify\DataRepository\Base;
+
+use Doctrine\DBAL\Connection;
+use Maatify\Common\Contracts\Adapter\AdapterInterface;
+use PDO;
+
+abstract class BaseMySQLRepository extends BaseRepository
+{
+    public function __construct(AdapterInterface $adapter)
+    {
+        parent::__construct($adapter);
+    }
+
+    protected function getPdoConnection(): PDO
+    {
+        $driver = $this->assertDriverAvailable(
+            $this->adapter->getDriver(),
+            'PDO'
+        );
+
+        return $this->assertDriverInstance($driver, PDO::class);
+    }
+
+    protected function getDbalConnection(): Connection
+    {
+        $driver = $this->assertDriverAvailable(
+            $this->adapter->getDriver(),
+            'Doctrine DBAL'
+        );
+
+        return $this->assertDriverInstance($driver, Connection::class);
+    }
+}